Move steven_shared into steven_protocol
