+++ Cargo.toml
@@ -61,7 +61,3 @@ version = "0"
 [dependencies.steven_protocol]
 path = "./protocol"
 version = "0"
-
-[dependencies.steven_shared]
-path = "./shared"
-version = "0"
