+++ protocol/Cargo.toml
@@ -16,9 +16,5 @@ log = { version = "0.4.6", features = ["std"] }
 flate2 = { version = "1.0.7", features = ["rust_backend"], default-features = false }
 num-traits = "0.2.8"
 
-[dependencies.steven_shared]
-path = "../shared"
-version = "0"
-
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 reqwest = "0.9.17"
