+++ protocol/src/direction.rs
