--- shared/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "steven_shared"
-version = "0.0.1"
-authors = [ "Thinkofdeath <thinkofdeath@spigotmc.org>" ]
-edition = "2018"
-
-[dependencies]
