--- shared/src/lib.rs
@@ -1,9 +1,0 @@
-
-pub mod axis;
-pub use self::axis::Axis;
-
-pub mod direction;
-pub use self::direction::Direction;
-
-pub mod position;
-pub use self::position::Position;
